refactor(pages): moved nested pages to pages main directory
