+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { SocketContext, socket } from './context/Context';
 
 // Components
-import Login from './pages/Login/Login';
+import Login from './pages/Login';
 import RoutesComponent from './components/Navigation/Routes';
 import { useEffect } from 'react';
