+++ client/src/pages/Searching.jsx
@@ -1,8 +1,8 @@
 import { useState, useContext, useEffect } from 'react';
 import { useSelector } from 'react-redux';
-import { SocketContext } from '../../context/Context';
+import { SocketContext } from '../context/Context';
 
-import Anonymous from '../Anonymous';
+import Anonymous from './Anonymous';
 
 const Searching = () => {
     // eslint-disable-next-line no-unused-vars
