+++ client/src/pages/Start.jsx
